requirements.txt
# Lead Agent - Python Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7

# HTTP Requests
httpx==0.25.1
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.24.3

# AI/ML Libraries
openai==1.3.7
langchain==0.0.350
transformers==4.35.2

# Email Services
smtplib2==0.2.1
email-validator==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# CORS
fastapi-cors==0.0.6

# Data Validation
pydantic==2.5.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Scraping & Lead Generation
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0

# API Rate Limiting
slowapi==0.1.9

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0


anthropic==0.7.8
